AVL, Splay and RedBlack Tree data strucutures
AVL, Splay and RB trees provide an upper limit of O(logn)[expect Splay which provides a worst case of O(n) but has a O(logn) expected). This is in contrast to Hash maps, where a bad hash function may slow the performance drastically. Here, we know these trees work regardless of the sequence pattern. Splay trees are useful if we keys appear more often than the others. Red black trees are useful if there is an upper bound on performance.